refactor: update operation fail error message
