+++ src/constants.js/errorMessage.js
@@ -1,0 +1,1 @@
+export const ERROR_MESSAGE = 'Operation failed';
