+++ src/nwd/checkDirentExist.js
